reading: change value to bigint

+++ lib/buzzn/schemas/constraints/reading/single.rb
@@ -2,8 +2,8 @@ require './app/models/reading/single.rb'
 require_relative '../reading'
 
 Schemas::Constraints::Reading::Single = Schemas::Support.Form do
-  required(:raw_value).filled(:int?)
-  required(:value).filled(:int?)
+  required(:raw_value).filled(:bigint?)
+  required(:value).filled(:bigint?)
   required(:unit).value(included_in?: Reading::Single.units.values)
   required(:reason).value(included_in?: Reading::Single.reasons.values)
   required(:read_by).value(included_in?: Reading::Single.read_by.values)
